Fix circular reference issue
